dropping precision when rounding to int
we had encountered a case where the amount on the acquirer (Paysera) had
a different amount than in a transaction (by 1 cent).

This stringified amount is actually used for posting the amount and the
final int rounding always rounds down so if the division by
currency.rounding lands on the lower side the whole amount differes by
the smallest currency amount. I.e.:
```
amount = str(int(316.46 / 0.01))  # 31645, while it should be 31646
```